separated the slider code into a separate file and made the slider functions globally visible
